added gravity effects, working towards collision handling

+++ CppProj/Engine/Engine.cpp
@@ -5,7 +5,7 @@
 int Engine::SCREEN_WIDTH = 1600;
 int Engine::SCREEN_HEIGHT = 900;
 GLFWwindow* Engine::window = NULL;
-double Engine::dt = 0;
+float Engine::dt = 0;
 
 Engine::Engine() {
 
@@ -57,7 +57,7 @@ bool Engine::Initialize(const char* windowTitle) {
 	glEnable(GL_BLEND);
 	glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
 
-	lastTime = glfwGetTime();
+	lastTime = (float)glfwGetTime();
 
 	return true;
 
@@ -80,6 +80,6 @@ void Engine::EndRender() {
 	glfwSwapBuffers(window);
 }
 
-double Engine::GetDT() {
+float Engine::GetDT() {
 	return dt;
 }
